Split up to_pil and from_pil into 2 how-tos
